ZiiPOSQRGenerator/requirements.txt
Pillow
qrcode
pyinstaller
tkinter
glob
